% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs.R
\docType{data}
\name{frs}
\alias{frs}
\title{EPA Facility Registry System table of regulated sites}
\description{
This is a 2021-09 snapshot version of EPA's FRS.

You can look up sites by REGISTRY_ID, 
but to look up by NAICS_CODES you need the long file, which has 
1 row per site-NAICS pair (multiple rows for a site in multiple NAICS).
That file was called facdata 

The 2016 version had only 1,129,481 rows (facilities), 22 columns.
The 2021-09 version has 3.1 million rows, and 
retained only 9 columns for this package.
  \preformatted{

  Classes ‘data.table’ and 'data.frame':	3144206 obs. of  9 variables:
  
  $ REGISTRY_ID   : chr  "110058283236" "110038070002" "110070215939" "110015787683" ...
  $ PRIMARY_NAME  : chr  "ADAK COD COOPERATIVE LLC" "ADAK NAVAL AIR STATION SITE"
  $ SITE_TYPE_NAME: chr  "STATIONARY" "STATIONARY" "STATIONARY" "STATIONARY" ...
  
  $ PGM_SYS_ACRNMS: chr  "NPDES:AKU000292" "BRAC:AK4170024323, EIS:10606011"
  $ INTEREST_TYPES: chr  "ICIS-NPDES UNPERMITTED"
  $ NAICS_CODES   : chr  NA NA NA "928110" ...
  $ SIC_CODES     : chr  "2091" "4952" "5171" "9711, OWNE" ...
  
  $ lat           : num  51.9 51.9 51.9 51.9 51.9 ...
  $ lon           : num  -177 -177 -177 -177 -177 ...
  }
}
