% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_clean.R
\name{frs_clean}
\alias{frs_clean}
\title{clean Facility Registry System dataset}
\usage{
frs_clean(
  frs,
  usefulcolumns = c("LATITUDE83", "LONGITUDE83", "REGISTRY_ID", "PRIMARY_NAME",
    "NAICS_CODES", "PGM_SYS_ACRNMS")
)
}
\arguments{
\item{frs}{output of frs_read}
}
\description{
This is just a helper function used to create the dataset saved as data for use in EJAM
}
\seealso{
\link{frs_get} which uses \link{frs_download} \link{frs_unzip} \link{frs_read} \link{frs_clean}
}
