% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_download.R
\name{frs_download}
\alias{frs_download}
\title{download Facility Registry System dataset}
\usage{
frs_download(
  folder = ".",
  zfile = "national_single.zip",
  zipbaseurl = "https://ofmext.epa.gov/FLA/www3/state_files/"
)
}
\arguments{
\item{folder}{path}

\item{zfile}{filename}

\item{zipbaseurl}{url}
}
\description{
This is just a helper function used to create the dataset saved as data for use in EJAM
}
\seealso{
\link{frs_get} which uses \link{frs_download} \link{frs_unzip} \link{frs_read} \link{frs_clean}
}
