% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_get.R
\name{frs_get}
\alias{frs_get}
\title{download, unzip, read, clean the Facility Registry System dataset}
\usage{
frs_get(
  folder = ".",
  date = Sys.Date(),
  zfile = "national_single.zip",
  zipbaseurl = "https://ofmext.epa.gov/FLA/www3/state_files/",
  csvfile = "NATIONAL_SINGLE.CSV",
  downloaded = FALSE
)
}
\arguments{
\item{folder}{path}

\item{date}{default is Sys.Date() which is today, but this is used as an attribute assigned to the results, 
representing the vintage, such as the date the frs was downloaded, obtained.}

\item{zfile}{name of zipfile}

\item{zipbaseurl}{url (without zipfile)}

\item{csvfile}{local filename from inside zip}

\item{downloaded}{default is FALSE, but can set to TRUE if previously downloaded the zipfile and now just need to unzip and read it.}
}
\description{
This is just a helper function used to create the dataset saved as data for use in EJAM.
  It invisibly returns the table of data as a data.table.
}
\details{
\preformatted{

 As of 2022-01-31:
 Only 1/4 have both location and some industry code:  27% have location and industry code.
 Only 1/5 to 1/3 have NAICS or at least SIC.
 2/3 have location.
 2/3 lack industry code, 1/3 lack location.
 
 67 percent HAVE NEITHER NAICS NOR SIC CODES.
 11 percent have both NAICS and SIC, 
 9.5 percent have just NAICS = 
    21 percent have NAICS. 
 12.5 percent have just SIC. 
 
  
  A large share of all the FRS sites (almost a third) have no latitude or longitude
  Those with lat lon have problems also:
  Some are are not in the USA.
  Some have errors in country code.
  Some use alternate ways of specifying USA.
}
}
\seealso{
\link{frs_get} which uses \link{frs_download} \link{frs_unzip} \link{frs_read} \link{frs_clean}
}
