% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_read.R
\name{frs_read}
\alias{frs_read}
\title{read Facility Registry System dataset}
\usage{
frs_read(csvfile = "NATIONAL_SINGLE.CSV", get_only_essential_cols = FALSE, ...)
}
\arguments{
\item{csvfile}{output of frs_unzip}

\item{get_only_essential_cols}{whether to keep only a few columns needed for EJAM package}

\item{...}{passed to readr::read_csv}
}
\description{
This is just a helper function used to create the dataset saved as data for use in EJAM
 Using readr::read_csv here this way is extremely slow, for some reason. million of rows.
}
\details{
\preformatted{

Source code for \link{frs_get} has more details on which fields might be useful.
 
 
}
}
\seealso{
\link{frs_get} which uses \link{frs_download} \link{frs_unzip} \link{frs_read} \link{frs_clean}
}
