% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frs_unzip.R
\name{frs_unzip}
\alias{frs_unzip}
\title{unzip Facility Registry System dataset}
\usage{
frs_unzip(zfile = "national_single.zip", ...)
}
\arguments{
\item{zfile}{zipfile obtained via frs_download}

\item{...}{passed to unzip}
}
\description{
This is just a helper function used to create the dataset saved as data for use in EJAM
}
\seealso{
\link{frs_get} which uses \link{frs_download} \link{frs_unzip} \link{frs_read} \link{frs_clean}
}
